Added Lesson 3 - Introduction to Classes
It is at this point that we really began to ramp up the difficulty. Classes and their interactions, when mastered, are a powerful tool for coding complex project neatly. Notice how things begin to become less concrete; variables are being used more than constants because they allow for code to respond to change.